# perf: use CourseOverview instead of modulestore when getting course summaries