--- a/cms/djangoapps/contentstore/views/course.py
+++ b/cms/djangoapps/contentstore/views/course.py
@@ -489,7 +489,7 @@ def _accessible_courses_list_from_groups(request):
     course_keys = list(course_keys.values())
 
     if course_keys:
-        courses_list = modulestore().get_course_summaries(course_keys=course_keys)
+        courses_list = CourseOverview.get_all_courses(filter_={'id__in': course_keys})
 
     return courses_list, []