## Description

This PR improves studio course listing performance for users with course access roles like staff or instructor by changing from MongoDB queries to Mysql queries. 

We measured the improvements as follows:

After 100 requests:

| Number of courses | Avg time elapsed before (s)  | Avg time elapsed after (s) |
|---------------------------|---------------------------------------|------------------------------------|
|         50                   |                2.34                        |         2.28                           |
|         107                 |                3.16                        |         2.57                           |
|         548                 |                7.91                        |         3.16                           |
|         1069               |                Error 500                |         4.62                           |
|         3380               |                Error 500                |         8.21                           |


If you go to /home using users with access roles staff/instructor in an org with more than ~500 courses, the page will collapse and raise a Network Timeout avoiding users to create new courses from the home page.

**Versions**
Mongo 4.0
MySQL 5.7
Lilac release

## Testing instructions

Go to Studio Home with a user with access roles (instructor/staff) over an organization, then the list of courses must match the CourseOverviews.

## Rejected strategies

We first tried to paginate the course listing (1st attempt in this [PR](https://github.com/openedx/edx-platform/pull/29902/)) but then realized we had to paginate both the course overview queryset and the mongo cursor. After trying with the cursor, we dropped the idea, given all the changes needed to make it work. Has this been tried before? I found [this](https://openedx.atlassian.net/browse/EDUCATOR-1258) ticket, but no implementation info.

## Concerns
According to [this](https://discuss.openedx.org/t/adr-for-removing-mongodb-from-edx-platform/6001), this change won't be necessary after altogether deprecating MongoDB from edxapp. There are some PR(s) linked on the post, but according to [these lines](https://github.com/openedx/edx-platform/pull/29058/files#diff-8eb3411fa751186021d817958f3bad09004d6a6e18d925f909677cba6b01c493R604-R647), we are still reading from MongoDB so that these changes could be helpful for organizations with numerous courses.