@mariajgrimaldi 2022-02-17
Hi community! I have a question, do you know what's the difference between `CourseOverview.get_all_courses` and `modulestore().get_course_summaries()` in [these](https://github.com/openedx/edx-platform/blob/master/cms/djangoapps/contentstore/views/course.py#L405-L409) lines?