@ormsbee 2022-02-18
> Hi community! I have a question, do you know what's the difference between CourseOverview.get_all_courses and modulestore().get_course_summaries() in [these](https://github.com/openedx/edx-platform/blob/master/cms/djangoapps/contentstore/views/course.py#L405-L409) lines?

The `get_course_summaries` call is always going to hit MongoDB, while the `get_all_courses` call will avoid MongoDB most of the time. They return different objects, which is always a little scary, but it seems like we already mix those in other combinations, so that makes me less concerned. [CourseSummary](https://github.com/openedx/edx-platform/blob/db32ff2cdf678fa8edd12c9da76a76eef0478614/common/lib/xmodule/xmodule/course_module.py#L1555) only has a small handful of fields declared on it. Honestly, I don't know if there's any value in ever calling CourseSummary these days.

One weird thing that you'll bump into is case sensitivity differences, but since CourseOverview is already used in other conditional branches for rendering the list of courses in Studio, I think that's fine.