@doctoryes 2022-02-18
@mariajgrimaldi You should rebase this PR on master to include the changes in this recently-merged PR:
https://github.com/openedx/edx-platform/pull/29950
The change will hopefully fix your test failures.