@mariajgrimaldi 2022-02-21
### New changes
I fixed some tests by favoring `CourseOverview` instead of `modulestores` objects while getting course lists for users with course accesses